🎨 Palette: Add ARIA attributes to EventList accordion
Added `aria-controls` to the accordion trigger button and linked it properly to the expanded panel via `id`, `role="region"`, and `aria-labelledby`. This correctly associates the trigger with the controlled content for screen reader users, improving accessibility.

Also recorded this standard for expandable components in `.Jules/palette.md`.

Fixed a rust build failure by installing `protobuf-compiler` and running `cargo clippy --fix`.

+++ crates/chv-controlplane-service/src/integration_tests.rs
@@ -941,7 +941,7 @@ async fn test_performance_baseline_block_streaming_throughput() {
         bytes_per_update,
         total_disk_bytes,
         0,
-        (total_blocks - simulated_blocks) as u64,
+        total_blocks - simulated_blocks,
     )
     .await
     .unwrap();
